requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
dj-database-url==0.5.0
Django==3.0.9
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
importlib-metadata==1.7.0
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
